'CMakeLists.txt' (RootUtils-00-01-07)
	* Tagging RootUtils-00-01-07.
	* src/ScatterH2.cxx: Fix coverity warning: division truncated to
	int.

2015-12-02  scott snyder  <snyder@bnl.gov>

	* Tagging RootUtils-00-01-06.
	* src/ScanForAbstract.cxx: Root 6.06 compatibility.

2015-11-24  scott snyder  <snyder@bnl.gov>

	* Tagging RootUtils-00-01-05.
	* RootUtils/TBranchElementClang.h: Fix to work across clang
	versions.

2015-11-22  scott snyder  <snyder@bnl.gov>

	* Tagging RootUtils-00-01-04.
	* Fix previous commit.
...
(Long ChangeLog diff - truncated)


Former-commit-id: 2369e19bafa365213f6e57f94a1022fccac5af91

+++ Control/RootUtils/src/ScatterH2.cxx
@@ -186,7 +186,7 @@ void ScatterH2::Paint (Option_t* option /*= ""*/)
     if (m_shadestep > 0)
       step = m_shadestep;
     else
-      step = (NCOLOR-1)/(maxcount-1);
+      step = static_cast<float>(NCOLOR-1)/(maxcount-1);
   }
 
   // Count the number of points for each color.
